+++ CommonplaceBookApp/NotebookStructureViewController.swift
@@ -26,6 +26,7 @@ protocol NotebookStructureViewControllerDelegate: AnyObject {
 
 /// Displays a list of any "structure" inside the notebook -- currently just hashtags
 final class NotebookStructureViewController: UIViewController {
+  /// What subset of notebook pages does the person want to see?
   enum StructureIdentifier: Hashable, CustomStringConvertible {
     case allNotes
     case hashtag(String)
@@ -33,14 +34,24 @@ final class NotebookStructureViewController: UIViewController {
     var description: String {
       switch self {
       case .allNotes: return "All Notes"
-      case .hashtag(let hashtag): return hashtag
+      case .hashtag(let hashtag): return String(hashtag.split(separator: "/").last!)
       }
     }
   }
 
+  /// Sections of our list.
   private enum Section: CaseIterable {
     case allNotes
     case hashtags
+  }
+
+  /// Item identifier -- this is separate from StructureIdentifier because we need to know if we have children for the disclosure indicator
+  private struct Item: Hashable, CustomStringConvertible {
+    var structureIdentifier: StructureIdentifier
+    var hasChildren: Bool
+
+    var description: String { structureIdentifier.description }
+    static let allNotes = Item(structureIdentifier: .allNotes, hasChildren: false)
   }
 
   public init(database: NoteDatabase) {
@@ -74,14 +85,22 @@ final class NotebookStructureViewController: UIViewController {
     return view
   }()
 
-  private lazy var dataSource: UICollectionViewDiffableDataSource<Section, StructureIdentifier> = {
-    let hashtagRegistration = UICollectionView.CellRegistration<UICollectionViewListCell, StructureIdentifier> { cell, _, structureIdentifier in
+  private lazy var dataSource: UICollectionViewDiffableDataSource<Section, Item> = {
+    let hashtagRegistration = UICollectionView.CellRegistration<UICollectionViewListCell, Item> { cell, _, item in
       var contentConfiguration = cell.defaultContentConfiguration()
-      contentConfiguration.text = structureIdentifier.description
+      contentConfiguration.text = item.description
       contentConfiguration.textProperties.color = .label
       cell.contentConfiguration = contentConfiguration
+
+      // Only items with children get an outline disclosure identifier.
+      if item.hasChildren {
+        cell.accessories = [.outlineDisclosure()]
+      } else {
+        cell.accessories = []
+      }
     }
-    let dataSource = UICollectionViewDiffableDataSource<Section, StructureIdentifier>(collectionView: collectionView) { (view, indexPath, item) -> UICollectionViewCell? in
+
+    let dataSource = UICollectionViewDiffableDataSource<Section, Item>(collectionView: collectionView) { (view, indexPath, item) -> UICollectionViewCell? in
       view.dequeueConfiguredReusableCell(using: hashtagRegistration, for: indexPath, item: item)
     }
 
@@ -91,7 +110,6 @@ final class NotebookStructureViewController: UIViewController {
         Logger.shared.debug("Skipping supplementary view for \(indexPath) because it isn't the hashtag section")
         return nil
       }
-      Logger.shared.info("This is the hashtag section, getting its header and footer")
       if kind == UICollectionView.elementKindSectionHeader {
         let headerRegistration = UICollectionView.SupplementaryRegistration<UICollectionViewCell>(elementKind: UICollectionView.elementKindSectionHeader) { headerView, _, _ in
           var headerConfiguration = UIListContentConfiguration.sidebarHeader()
@@ -146,9 +164,47 @@ final class NotebookStructureViewController: UIViewController {
 
 extension NotebookStructureViewController: UICollectionViewDelegate {
   func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-    if let structureIdentifier = dataSource.itemIdentifier(for: indexPath) {
-      delegate?.notebookStructureViewController(self, didSelect: structureIdentifier)
+    if let item = dataSource.itemIdentifier(for: indexPath) {
+      delegate?.notebookStructureViewController(self, didSelect: item.structureIdentifier)
       splitViewController?.show(.supplementary)
+    }
+  }
+
+  func collectionView(_ collectionView: UICollectionView, contextMenuConfigurationForItemAt indexPath: IndexPath, point: CGPoint) -> UIContextMenuConfiguration? {
+    let item = dataSource.itemIdentifier(for: indexPath)
+    guard case .hashtag(let hashtag) = item?.structureIdentifier else { return nil }
+    return UIContextMenuConfiguration(identifier: nil, previewProvider: nil) { [weak self, database] _ in
+      guard let self = self else { return nil }
+      let rename = UIAction(title: "Rename \(hashtag)", image: UIImage(systemName: "square.and.pencil")) { _ in
+        Logger.shared.debug("Rename \(hashtag)")
+        let alert = UIAlertController(title: "Rename \(hashtag)", message: nil, preferredStyle: .alert)
+        alert.addTextField { textField in
+          textField.placeholder = "#new-hashtag"
+        }
+        alert.addAction(UIAlertAction(title: "OK", style: .default, handler: { _ in
+          guard var newHashtag = alert.textFields?.first?.text else {
+            Logger.shared.info("No replacement text; skipping hashtag rename.")
+            return
+          }
+          if newHashtag.first != "#" {
+            newHashtag = "#\(newHashtag)"
+          }
+          Logger.shared.info("Replacing \(hashtag) with \(newHashtag)")
+          do {
+            try database.renameHashtag(
+              hashtag,
+              to: newHashtag,
+              filter: { $0.hashtags.anySatisfy { noteHashtag in hashtag.isPathPrefix(of: noteHashtag) } }
+            )
+          } catch {
+            Logger.shared.error("Error renaming hashtag: \(error)")
+          }
+        }))
+        alert.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: nil))
+        alert.view.tintColor = .grailTint
+        self.present(alert, animated: true, completion: nil)
+      }
+      return UIMenu(title: "", children: [rename])
     }
   }
 }
@@ -165,13 +221,47 @@ private extension NotebookStructureViewController {
   }
 
   func updateSnapshot() {
-    var snapshot = NSDiffableDataSourceSnapshot<Section, StructureIdentifier>()
+    var snapshot = NSDiffableDataSourceSnapshot<Section, Item>()
     snapshot.appendSections([.allNotes])
     snapshot.appendItems([.allNotes])
-    if !database.hashtags.isEmpty {
+    let hashtagSectionSnapshot = makeHashtagSectionSnapshot()
+    if !hashtagSectionSnapshot.items.isEmpty {
       snapshot.appendSections([.hashtags])
-      snapshot.appendItems(database.hashtags.map { .hashtag($0) })
     }
     dataSource.apply(snapshot)
+    if !hashtagSectionSnapshot.items.isEmpty {
+      dataSource.apply(hashtagSectionSnapshot, to: .hashtags)
+    }
+  }
+
+  private func makeHashtagSectionSnapshot() -> NSDiffableDataSourceSectionSnapshot<Item> {
+    var snapshot = NSDiffableDataSourceSectionSnapshot<Item>()
+
+    // Enumerate every hashtag and make sure there is an entry for the hashtag *and all prefixes*.
+    // Denote that each prefix has children.
+    //
+    // This algorithm depends on database.hashtags being sorted, so we will never process a prefix *after* a more specific
+    // string. E.g., things work if we process `#books` then `#books/2020`, but will break if we process `#books/2020` before
+    // `#books`.
+    var stringToItem = [String: Item]()
+    for hashtag in database.hashtags {
+      for (index, character) in hashtag.enumerated() where character == "/" {
+        let prefix = String(hashtag.prefix(index))
+        stringToItem[prefix] = Item(structureIdentifier: .hashtag(prefix), hasChildren: true)
+      }
+      stringToItem[hashtag] = Item(structureIdentifier: .hashtag(hashtag), hasChildren: false)
+    }
+
+    // Now make a snapshot item for everything in `stringToItem`. The only tricky part is for a multi-part
+    // hashtag like `#books/2020`, we have to look up its parent `#books`... and again, this depends on the array
+    // being sorted.
+    for hashtag in stringToItem.keys.sorted() {
+      let parent = hashtag
+        .lastIndex(of: "/")
+        .flatMap { String(hashtag.prefix(upTo: $0)) }
+        .flatMap { stringToItem[$0] }
+      snapshot.append([stringToItem[hashtag]!], to: parent)
+    }
+    return snapshot
   }
 }